WIP: refactor subcommands and add first template for the template-list command
* changes 'template list' and 'template show' subcommands to template-list and template-show as first class commands
* removes aggregate command class and references to it
* adds a template for the template-list command using the new markdown templates/greenbar
