+++ Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'cog-rb', '~> 0.3.0'
+gem 'cog-rb', github: 'operable/cog-rb', branch: 'imbriaco/remove-subcommands'
 gem 'aws-sdk', '~> 2.5'
 
 group "development", "test" do
