+++ templates/template_table.md
@@ -1,0 +1,4 @@
+| Name | Last Modified |
+| ---- | ------------- |
+~each var=$results as=template~| ~$template.name~ | ~$template.last_modified~ |
+~end~
